Move val mapping to its own submodule and document
