+++ qcodes/instrument_drivers/Keysight/Infiniium.py
@@ -11,8 +11,7 @@ from qcodes.instrument import VisaInstrument
 from qcodes.instrument.base import InstrumentBase
 from qcodes.instrument.channel import ChannelList, InstrumentChannel, InstrumentModule
 from qcodes.parameters import Parameter, ParameterWithSetpoints
-from qcodes.utils import deprecate
-from qcodes.utils.helpers import create_on_off_val_mapping
+from qcodes.utils import create_on_off_val_mapping, deprecate
 
 
 class DSOTimeAxisParam(Parameter):
