+++ qcodes/instrument_drivers/Keysight/KtM960x.py
@@ -5,7 +5,7 @@ from typing import Any, Dict, Optional, Tuple
 from qcodes import validators as vals
 from qcodes.instrument import Instrument
 from qcodes.parameters import MultiParameter, ParamRawDataType
-from qcodes.utils.helpers import create_on_off_val_mapping
+from qcodes.utils import create_on_off_val_mapping
 
 from .KtM960xDefs import *
