+++ qcodes/instrument_drivers/Keysight/N9030B.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from qcodes.instrument import InstrumentChannel, VisaInstrument
 from qcodes.parameters import Parameter, ParameterWithSetpoints, ParamRawDataType
-from qcodes.utils.helpers import create_on_off_val_mapping
+from qcodes.utils import create_on_off_val_mapping
 from qcodes.validators import Arrays, Enum, Ints, Numbers
