+++ qcodes/instrument_drivers/agilent/E8527D.py
@@ -5,7 +5,7 @@ from numpy import pi
 
 import qcodes.validators as vals
 from qcodes.instrument import VisaInstrument
-from qcodes.utils.helpers import create_on_off_val_mapping
+from qcodes.utils import create_on_off_val_mapping
 
 # todo should this be public
 from qcodes.validators.validators import numbertypes
