+++ qcodes/instrument_drivers/tektronix/DPO7200xx.py
@@ -12,7 +12,7 @@ import numpy as np
 
 from qcodes.instrument import ChannelList, Instrument, InstrumentChannel, VisaInstrument
 from qcodes.parameters import Parameter, ParameterWithSetpoints
-from qcodes.utils.helpers import create_on_off_val_mapping
+from qcodes.utils import create_on_off_val_mapping
 from qcodes.validators import Arrays, Enum
