+++ qcodes/instrument_drivers/tektronix/Keithley_2400.py
@@ -1,7 +1,7 @@
 from typing import Any
 
 from qcodes.instrument import VisaInstrument
-from qcodes.utils.helpers import create_on_off_val_mapping
+from qcodes.utils import create_on_off_val_mapping
 from qcodes.validators import Enum, Strings
