+++ qcodes/instrument_drivers/tektronix/Keithley_2600_channels.py
@@ -16,7 +16,7 @@ from qcodes.parameters import (
     ParameterWithSetpoints,
     ParamRawDataType,
 )
-from qcodes.utils.helpers import create_on_off_val_mapping
+from qcodes.utils import create_on_off_val_mapping
 
 log = logging.getLogger(__name__)
