+++ qcodes/instrument_drivers/tektronix/Keithley_3706A.py
@@ -5,7 +5,7 @@ from typing import List, Dict, Optional, Union, Tuple, Any
 
 import qcodes.validators as vals
 from qcodes.instrument import VisaInstrument
-from qcodes.utils.helpers import create_on_off_val_mapping
+from qcodes.utils import create_on_off_val_mapping
 
 
 class UnknownOrEmptySlot(Exception):
