+++ qcodes/instrument_drivers/tektronix/keithley_7510.py
@@ -11,7 +11,7 @@ from qcodes.parameters import (
     ParamRawDataType,
     invert_val_mapping,
 )
-from qcodes.utils.helpers import create_on_off_val_mapping
+from qcodes.utils import create_on_off_val_mapping
 from qcodes.validators import Arrays, Enum, Ints, Lists, Numbers
