+++ qcodes/tests/parameter/test_val_mapping.py
@@ -3,7 +3,7 @@ import pytest
 import qcodes.validators as vals
 from qcodes.parameters import Parameter
 from qcodes.tests.instrument_mocks import DummyInstrument
-from qcodes.utils.helpers import create_on_off_val_mapping
+from qcodes.utils import create_on_off_val_mapping
 
 from .conftest import ParameterMemory
